fix: whitelist middleware and release: client 0.6.4
