+++ client-npm/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@fdm-monster/client",
   "author": "David Zwart",
-  "version": "0.6.3",
+  "version": "0.6.4",
   "private": false,
   "license": "AGPL-3.0-or-later",
   "repository": {
