+++ client-vue/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@fdm-monster/client",
-  "version": "0.6.3",
+  "version": "0.6.4",
   "private": false,
   "author": "David Zwart",
   "scripts": {
