+++ client-vue/src/components/Settings/FdmSettings.vue
@@ -15,12 +15,11 @@
               <v-icon>info</v-icon> &nbsp; Be cautious, setting the wrong whitelist could make you
               lose access to the server!
             </v-alert>
-            Only allow access from specific IP Adresses or subnets. <br />
-            Examples:
+            Only allow access from specific IP Adresses or subnets. Note: 127.0.0.1 will always be
+            allowed access. Examples:
+            <br />
             <v-chip small>192.168</v-chip>
-            ,
             <v-chip small>192.168.1</v-chip>
-            or
             <v-chip small>192.168.1.1</v-chip>
             <br />
             <v-row>
@@ -56,7 +55,9 @@
             </v-row>
             <v-row>
               <v-col>
-                <v-btn color="default" @click="setWhitelistSettings()"> reset to default </v-btn>
+                <v-btn color="default" @click="resetWhitelistSettingsToDefault()">
+                  reset to default</v-btn
+                >
                 <v-btn color="primary" @click="setWhitelistSettings()">
                   save whitelist settings
                 </v-btn>
@@ -171,19 +172,25 @@ export default defineComponent({
       );
     },
     appendIpAddress(ip: string) {
+      if (!isValidIPOrMask(ip)) return;
       this.whitelistedIpAddresses.push(ip.toLowerCase());
+    },
+    async resetWhitelistSettingsToDefault() {
+      this.whitelistedIpAddresses = ["127.0.0.1", "::12"];
+      this.whitelistEnabled = false;
+      await this.setWhitelistSettings();
     },
     async setWhitelistSettings() {
       const serverSettings = await SettingsService.setWhitelistSettings({
         whitelistedIpAddresses: this.whitelistedIpAddresses,
         whitelistEnabled: this.whitelistEnabled,
       });
-      this.fileHandlingSettings = serverSettings.printerFileClean;
+      this.whitelistedIpAddresses = serverSettings.server?.whitelistedIpAddresses;
+      this.whitelistEnabled = serverSettings.server?.whitelistEnabled;
     },
     async setFileCleanSettings() {
       const serverSettings = await SettingsService.setFileCleanSettings(this.fileHandlingSettings);
-      this.whitelistedIpAddresses = serverSettings.server?.whitelistedIpAddresses;
-      this.whitelistEnabled = serverSettings.server?.whitelistEnabled;
+      this.fileHandlingSettings = serverSettings.printerFileClean;
     },
     async purgeFiles() {
       await PrinterFileService.purgeFiles();
