+++ server/middleware/authenticate.js
@@ -1,6 +1,5 @@
 const { inject } = require("awilix-express");
 const { AuthorizationError, AuthenticationError } = require("../exceptions/runtime.exceptions");
-const { asValue } = require("awilix");
 const { serverSettingKey } = require("../constants/server-settings.constants");
 
 function authorizePermission(permission) {
@@ -19,27 +18,6 @@ function authorizePermission(permission) {
 }
 
 module.exports = {
-  validateWhitelistedIp: inject(({ settingsStore, loggerFactory }) => async (req, res, next) => {
-    const logger = loggerFactory("validateWhitelistedIp");
-    const serverSettings = settingsStore.getServerSettings();
-    if (
-      (serverSettings && !serverSettings[serverSettingKey]) ||
-      serverSettings[serverSettingKey]?.whitelistEnabled
-    ) {
-      next();
-    }
-
-    const whitelist = serverSettings[serverSettingKey].whitelistedIpAddresses;
-    const ipAddress = req.connection.remoteAddress;
-    // Empty whitelist is treated as disabled as well
-    if (whitelist?.length && !whitelist.includes(ipAddress)) {
-      logger.error("IP was not whitelisted ", req.connection.remoteAddress);
-      const err = new Error("Bad IP: " + req.connection.remoteAddress);
-      next(err);
-    }
-
-    next();
-  }),
   authenticate: () =>
     inject(({ settingsStore }) => async (req, res, next) => {
       const serverSettings = settingsStore.getServerSettings();
