+++ server/package.json
@@ -40,7 +40,7 @@
     "vue"
   ],
   "dependencies": {
-    "@fdm-monster/client": "0.6.3",
+    "@fdm-monster/client": "0.6.4",
     "@influxdata/influxdb-client": "^1.31.0",
     "awilix": "8.0.0",
     "awilix-express": "8.0.0",
